+++ src/com/alibaba/f2e/ycombo/CombinerException.java
@@ -1,0 +1,17 @@
+/**
+ * YCombo
+ * Copyright (c) 2012 Alibaba.com, Inc.
+ * MIT Licensed
+ * @author Nanqiao Deng
+ */
+
+package com.alibaba.f2e.ycombo;
+
+/**
+ * Class for combiner exception.
+ */
+public class CombinerException extends Exception {
+	public CombinerException(String message) {
+		super(message);
+	}
+}