+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 2012.02.13, Version 0.1.0
 -------------------------
-+ initial release
+* initial release
 
 2012.03.02, Version 0.1.1
 -------------------------
@@ -12,5 +12,11 @@
 
 2012.03.13, Version 0.1.3
 -------------------------
-* Change seed file extension name from seed.[TYPE] to [TYPE].seed to avoid conflict with Unicorn.
+* Change seed file extension name from seed.TYPE to TYPE.seed to avoid conflict with Unicorn.
 * Change common-line option "--prefix" to "--extname" to response to the change above.
+
+2012.03.18, Version 0.1.4
+-------------------------
+* Fix the log when input file has syntax error.
+* Error in one seed file will not break the whole combo task now.
+* Add new common-line option "--nocompress", allowing to combine file only.